Starting implementation of RealNetworker

+++ src/api/ComputeEngine/ComputeUserInput.java
@@ -2,18 +2,22 @@ package api;
 
 public class ComputeUserInput{
 	
-	private int value;
+	private long value;
 		
-	public ComputeUserInput(int userInput){		//change later?
-		 int value = userInput;
+	
+	
+	public ComputeUserInput(long userInput){		//change later?
+		 long value = userInput;
 	}
 	
 	public ComputeUserInput(){
-		 int value = 2;
+		 long value = 2;
 	}
 	
-	public int getValue(){
+	public long getValue(){
 		return this.value;
 	}
+	
+	
 	
 }
